+++ es/index.html
@@ -26,11 +26,12 @@
   <!-- App -->
   <link rel="stylesheet" href="/static/css/main.css"/>
   <link rel="stylesheet" href="/static/css/materialize.overrides.css"/>
-  <script type="application/javascript" src="/static/js/localization.es.js"></script>
-  <script type="module" src="/static/js/settings.js"></script>
+  <script type="application/javascript" src="/static/js/locales/es.js"></script>
+  <script type="module" src="/static/js/state.js"></script>
   <script type="module" src="/static/js/ui.js"></script>
   <script type="module" src="/static/js/plots.js"></script>
   <script type="module" src="/static/js/data.js"></script>
+  <script type="module" src="/static/js/map.js"></script>
   <script type="module" src="/static/js/main.js"></script>
 </head>
