+++ static/js/locales/en-US.js
