+++ static/js/locales/es.js
